requirements.txt
asgiref==3.8.1
django==4.2.14
django-user-agents==0.4.0
djangorestframework==3.15.2
pyyaml==6.0.1
sqlparse==0.5.0
ua-parser==0.18.0
user-agents==2.2.0
